requirements.txt
qrcode[pil]
pdf2image
Pillow